update to lasest version of openoakland:
the commit branches off of
https://github.com/code4sac/openbudgetsac.org/commit/4c5091109a8b3f7c44393d9fe02d54041e9e6af8,
where all of the content of 1.0 was deleted.

openoakland has added and deleted a lot
of code since we copied over the changes last year. they use 11ty
instead of harp, migrate most of their templates to pug, and they have a
github actions pipeline now

because the code in the release branch the openoakland repository do not
share a git history, i was unable to simply merge their changes on top
of ours. going to apply any changes that were made back on top of this
commit.